Fixes issues with kernel/extension compilation
This patch adds a __bareflank__ macro, similar to __linux__
that Clang will define, telling us that the code is being
cross-compiled for the Bareflank kernel and extensions.
This also remove the freestanding flag from the BSL as that
will need to be set by Clang as well when this macro is
defined.